# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "JobHunt OS API",
    "version": "0.1.0",
    "description": "Backend for JobHunt OS. Single-tenant.\n\n**Two invariants the UI must respect:**\n1. Every job carries `gaps` — JD requirements the candidate does *not* have. Always render them. Never let the UI imply the candidate has a gap technology.\n2. Nothing in this API sends anything to another human. Outbound actions (Phase 4) will be draft-and-approve only."
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "name": "x-api-key",
        "in": "header"
      }
    },
...
  },
  "paths": {
...
    "/facts": {
      "get": {
        "summary": "List verified facts from the fact table",
        "tags": [
          "facts"
        ],
        "description": "Returns the truth constraint facts used for screening answers and resume tailoring.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "personal",
                "contact",
                "authorization",
                "availability",
                "compensation",
                "experience",
                "education",
                "links",
                "preferences"
              ]
            },
            "in": "query",
            "name": "category",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "key": {
                            "type": "string"
                          },
                          "category": {
                            "type": "string",
                            "enum": [
                              "personal",
                              "contact",
                              "authorization",
                              "availability",
                              "compensation",
                              "experience",
                              "education",
                              "links",
                              "preferences"
                            ]
                          },
                          "label": {
                            "type": "string"
                          },
                          "value": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "number"
                              },
                              {
                                "type": "boolean"
                              },
                              {
                                "type": "array",
                                "items": {
                                  "type": "string"
                                }
                              }
                            ]
                          },
                          "evidence": {
                            "type": "string"
                          },
                          "verifiedAt": {
                            "type": "string"
                          },
                          "notes": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "key",
                          "category",
                          "label",
                          "value",
                          "verifiedAt"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "total": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "items",
                    "total"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create or update a verified fact",
        "tags": [
          "facts"
        ],
        "description": "Updates a fact or caches a newly confirmed answer from the Human Gate queue.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "key": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "personal",
                      "contact",
                      "authorization",
                      "availability",
                      "compensation",
                      "experience",
                      "education",
                      "links",
                      "preferences"
                    ]
                  },
                  "label": {
                    "type": "string"
                  },
                  "value": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "number"
                      },
                      {
                        "type": "boolean"
                      },
                      {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    ]
                  },
                  "evidence": {
                    "type": "string"
                  },
                  "verifiedAt": {
                    "type": "string"
                  },
                  "notes": {
                    "type": "string"
                  }
                },
                "required": [
                  "key",
                  "category",
                  "label",
                  "value",
                  "verifiedAt"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "key": {
                      "type": "string"
                    },
                    "category": {
                      "type": "string",
                      "enum": [
                        "personal",
                        "contact",
                        "authorization",
                        "availability",
                        "compensation",
                        "experience",
                        "education",
                        "links",
                        "preferences"
                      ]
                    },
                    "label": {
                      "type": "string"
                    },
                    "value": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "number"
                        },
                        {
                          "type": "boolean"
                        },
                        {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      ]
                    },
                    "evidence": {
                      "type": "string"
                    },
                    "verifiedAt": {
                      "type": "string"
                    },
                    "notes": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "key",
                    "category",
                    "label",
                    "value",
                    "verifiedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/facts/resolve": {
      "post": {
        "summary": "Resolve an ATS screening question",
        "tags": [
          "facts"
        ],
        "description": "Evaluates a screening question against the verified fact table. Returns answer (truthful), ask (queued for human), or abort (gap violation).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "question": {
                    "type": "string",
                    "minLength": 1
                  },
                  "options": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "question"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "outcome": {
                      "type": "string",
                      "enum": [
                        "answer",
                        "ask",
                        "abort"
                      ]
                    },
                    "key": {
                      "type": "string"
                    },
                    "answer": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "number"
                        },
                        {
                          "type": "boolean"
                        },
                        {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      ]
                    },
                    "formattedAnswer": {
                      "type": "string"
                    },
                    "confidence": {
                      "type": "number"
                    },
                    "reason": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "outcome",
                    "formattedAnswer",
                    "confidence",
                    "reason"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://127.0.0.1:4000"
    }
  ],
  "security": [],
...
}